// todoList-Frontend/src/common/typeInterface/Enum.tsx
export enum typesEnum {
  GET_POST = 'GET_POST',
  GET_POST_SUCCESS = 'GET_POST_SUCCESS',
  GET_POST_ERORR = 'GET_POST_ERORR',
  CREATE_POST = 'CREATE_POST',
  CREATE_POST_SUCCESS = 'CREATE_POST_SUCCESS',
  CREATE_POST_ERORR = 'CREATE_POST_ERORR',
  UPDATE_POST = 'UPDATE_POST',
  UPDATE_POST_SUCCESS = 'UPDATE_POST_SUCCESS',
  UPDATE_POST_ERORR = 'UPDATE_POST_ERORR',
  DELETE_POST = 'DELETE_POST',
  DELETE_POST_SUCCESS = 'DELETE_POST_SUCCESS',
  DELETE_POST_ERORR = 'DELETE_POST_ERORR',
  SHOWMODALUPDATE = 'SHOWMODALUPDATE',
  CLOSEMODALUPDATE = 'CLOSEMODALUPDATE',
  SHOWMODALCREATE = 'SHOWMODALCREATE',
  CLOSEMODALCREATE = 'CLOSEMODALCREATE',
  TOGGLEDARK = 'TOGGLEDARK',
}

export enum restApiEnum {
  GET = 'GET',
  POST = 'POST',
  PUT = 'PUT',
  DELETE = 'DELETE',
}